html/webappapis/timers/evil-spec-example.html

<!doctype html>
<title>Interaction of step_timeout and WebIDL</title>
<link rel="author" title="Ian Hickson" href="mailto:ian@hixie.ch">
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-windowtimers-settimeout">
<link rel="help" href="https://heycam.github.io/webidl/#es-operations">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var t = async_test()
function finishTest() {
  assert_equals(log, "ONE TWO ")
  t.done()
}
var log = '';
function logger(s) { log += s + ' '; }

step_timeout({ toString: function () {
  step_timeout("logger('ONE')", 100);
  return "logger('TWO'); t.step(finishTest)";
} }, 100);
</script>
